+++ podcasts/Sharing/ShareDestination.swift
@@ -1,6 +1,7 @@
 import SwiftUI
 import PocketCastsDataModel
 import Combine
+import PocketCastsUtils
 
 enum ShareDestination: Hashable {
     case instagram
@@ -75,8 +76,10 @@ enum ShareDestination: Hashable {
                 activityViewController.popoverPresentationController?.sourceRect = rect
             }
             activityViewController.completionWithItemsHandler = { activityType, completed, returnedItems, activityError in
+                NotificationCenter.postOnMainThread(notification: podcasts.Constants.Notifications.closedNonOverlayableWindow)
                 receiver.cancel()
             }
+            NotificationCenter.postOnMainThread(notification: podcasts.Constants.Notifications.openingNonOverlayableWindow)
             vc.presentedViewController?.present(activityViewController, animated: true, completion: {
                 ShareDestination.logClipShared(option: option, style: style, clipUUID: clipUUID, source: source)
                 ShareDestination.logPodcastShared(style: style, option: option, destination: self, source: source)